AppParamParser: relocate to application directory
