+++ xbmc/input/InputManager.cpp
@@ -9,7 +9,6 @@
 #include "InputManager.h"
 
 #include "AppInboundProtocol.h"
-#include "AppParamParser.h"
 #include "Application.h"
 #include "ButtonTranslator.h"
 #include "CustomControllerTranslator.h"
@@ -20,6 +19,7 @@
 #include "TouchTranslator.h"
 #include "Util.h"
 #include "XBMC_vkeys.h"
+#include "application/AppParamParser.h"
 #include "guilib/GUIAudioManager.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIControl.h"
