+++ xbmc/platform/darwin/osx/XBMCApplication.mm
@@ -10,8 +10,8 @@
 #import "XBMCApplication.h"
 
 #include "AppInboundProtocol.h"
-#include "AppParamParser.h"
 #include "ServiceBroker.h"
+#include "application/AppParamParser.h"
 #include "messaging/ApplicationMessenger.h"
 #include "platform/xbmc.h"
 #include "settings/AdvancedSettings.h"
