+++ xbmc/platform/linux/AppParamParserLinux.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "AppParamParser.h"
+#include "application/AppParamParser.h"
 
 class CAppParamParserLinux : public CAppParamParser
 {
