+++ xbmc/windowing/osx/OpenGL/OSXGLWindow.mm
@@ -9,9 +9,9 @@
 #import "OSXGLWindow.h"
 
 #include "AppInboundProtocol.h"
-#include "AppParamParser.h"
 #include "Application.h"
 #include "ServiceBroker.h"
+#include "application/AppParamParser.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
 #include "messaging/ApplicationMessenger.h"
